+++ dbc/src/lib.rs
@@ -27,12 +27,12 @@
 
 #[macro_use]
 extern crate amplify;
-//#[macro_use]
 #[cfg(feature = "miniscript")]
 extern crate miniscript_crate as miniscript;
 #[cfg(feature = "serde")]
-// #[macro_use]
+#[macro_use]
 extern crate serde_crate as serde;
+//#[macro_use]
 extern crate strict_encoding;
 
 pub mod tapret;
